refactoring: define LockFilter.
